requirements.txt
Django>=3.1.3,<3.1.4
djangorestframework>=3.12.2,<3.12.3
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
django-filter>=2.4.0
flake8>=3.6.0,<3.7.0